rename file print_formats, add 1 function in file
